refactor: make IdempotencyHandler generic for any fn
